test: refactor frames root tests (#8012)
