+++ e2e/nested-frame-navigation/.vscode/launch.json
@@ -22,7 +22,7 @@
                 "4723",
                 "-a",
                 "--grep",
-                "layout-root"
+                "layout-root-with-multi-frames"
             ],
             "internalConsoleOptions": "openOnSessionStart"
         },
