+++ e2e/nested-frame-navigation/e2e/config.ts
@@ -1,4 +1,4 @@
 export const suspendTime = 1;
 export const appSuspendResume = true;
-export const dontKeepActivities = true;
-export const transitions = ["Default", "None", "Slide", "Flip"];
+export const dontKeepActivities = process.env["DONT_KEEP_ACTIVITIES"] === "true" ? true : false;
+export const allTransitions = ["Default", "None", "Slide", "Flip"];
