+++ e2e/nested-frame-navigation/e2e/frame-root/tab-view.e2e-spec.ts
@@ -1,24 +1,36 @@
 import { AppiumDriver, createDriver, logWarn, nsCapabilities } from "nativescript-dev-appium";
 
-import { Screen, playersData, somePage, teamsData, driverDefaultWaitTime, Item } from "./screen";
-import * as shared from "./shared.e2e-spec";
-import { suspendTime, appSuspendResume, dontKeepActivities, transitions } from "./config";
-import { TabNavigationScreen } from "./tab-navigation-screen";
+import { Screen, playersData, somePage, teamsData, driverDefaultWaitTime, Item } from "../screens/screen";
+import * as shared from "../screens/shared";
+import { suspendTime, appSuspendResume, dontKeepActivities, allTransitions } from "../config";
+import { TabViewNavigationScreen } from "../screens/tabview-navigation-screen";
 
-const roots = ["TabsTop", "TabsBottom"];
+// NOTE: TabViewTop is Android only scenario (for iOS we will essentially execute 2x TabViewBottom)
+const roots = ["TabViewTop", "TabViewBottom"];
 
-const rootType = "frame-tabs-root";
-describe(rootType, async function () {
+describe("frame-root-with-tab-view", async function () {
     let driver: AppiumDriver;
     let screen: Screen;
+    let transitions = [...allTransitions];
 
     before(async function () {
         nsCapabilities.testReporter.context = this;
-        logWarn(`====== ${rootType} ========`);
+        logWarn(`====== frame-root-with-tab-view ========`);
         driver = await createDriver();
-        screen = new TabNavigationScreen(driver);
+        await driver.restartApp();
+        screen = new TabViewNavigationScreen(driver);
         if (dontKeepActivities) {
             await driver.setDontKeepActivities(true);
+        }
+        if (driver.isIOS) {
+            logWarn("NOTE: TabViewTop is Android only scenario (for iOS we will not execute it)");
+            roots.shift();
+        }
+
+        if (shared.isApiLevel19(driver)) {
+            // TODO: known issue https://github.com/NativeScript/NativeScript/issues/6798
+            console.log("Skipping flip transition tests on api level 19");
+            transitions = transitions.filter(tr => !tr.toLowerCase().includes("flip"));
         }
 
         driver.defaultWaitTime = driverDefaultWaitTime;
@@ -41,7 +53,7 @@ describe(rootType, async function () {
     for (let index = 0; index < roots.length; index++) {
         const root = roots[index];
 
-        describe(`${rootType}-${root} scenarios:`, async function () {
+        describe(`frame-root-with-tab-view-${root} scenarios:`, async function () {
             logWarn(`===== Root: ${root}`);
             for (let trIndex = 0; trIndex < transitions.length; trIndex++) {
                 const transition = transitions[trIndex];
@@ -50,18 +62,11 @@ describe(rootType, async function () {
                 const teamOne: Item = teamsData[`teamOne${transition}`];
                 const teamTwo: Item = teamsData[`teamTwo${transition}`];
 
-                describe(`${rootType}-${root}-transition-${transition}-scenarios:`, async function () {
+                describe(`frame-root-with-tab-view-${root}-transition-${transition}-scenario:`, async function () {
 
                     before(async function () {
                         nsCapabilities.testReporter.context = this;
                         logWarn(`========= ${root}-${transition} =========`);
-
-                        if (transition === "Flip" &&
-                            driver.isAndroid && parseInt(driver.platformVersion) === 19) {
-                            // TODO: known issue https://github.com/NativeScript/NativeScript/issues/6798
-                            console.log("skipping flip transition tests on api level 19");
-                            this.skip();
-                        }
                     });
 
                     it("loaded home page", async function () {
@@ -80,11 +85,12 @@ describe(rootType, async function () {
                     it("loaded player details and go back twice", async function () {
                         await shared.testPlayerNavigated(playerTwo, screen);
 
-                        if (appSuspendResume) {
-                            await driver.backgroundApp(suspendTime);
-                            await screen.loadedElement(playerTwo.name); // wait for player
+                        if (!shared.preventApplicationCrashCauesByAutomation(driver)) {
+                            if (appSuspendResume) {
+                                await driver.backgroundApp(suspendTime);
+                                await screen.loadedElement(playerTwo.name); // wait for player
+                            }
                         }
-
                         await shared.testPlayerNavigatedBack(screen, driver);
 
                         if (appSuspendResume) {
@@ -116,9 +122,11 @@ describe(rootType, async function () {
                     it("loaded player details and navigate parent frame and go back", async function () {
                         await shared.testPlayerNavigated(playerTwo, screen);
 
-                        if (appSuspendResume) {
-                            await driver.backgroundApp(suspendTime);
-                            await screen.loadedElement(playerTwo.name); // wait for player
+                        if (!shared.preventApplicationCrashCauesByAutomation(driver)) {
+                            if (appSuspendResume) {
+                                await driver.backgroundApp(suspendTime);
+                                await screen.loadedElement(playerTwo.name); // wait for player
+                            }
                         }
 
                         await shared[`testSomePageNavigated${transition}`](screen);
@@ -165,9 +173,11 @@ describe(rootType, async function () {
 
                         await shared.testPlayerNavigated(playerTwo, screen);
 
-                        if (appSuspendResume) {
-                            await driver.backgroundApp(suspendTime);
-                            await screen.loadedElement(playerTwo.name); // wait for player
+                        if (!shared.preventApplicationCrashCauesByAutomation(driver)) {
+                            if (appSuspendResume) {
+                                await driver.backgroundApp(suspendTime);
+                                await screen.loadedElement(playerTwo.name); // wait for player
+                            }
                         }
 
                         await screen.loadedPlayerDetails(playerTwo);
@@ -205,12 +215,13 @@ describe(rootType, async function () {
 
                         await shared.testTeamNavigated(teamTwo, screen);
 
-                        if (appSuspendResume) {
-                            await screen.loadedElement(teamTwo.name); // wait for team
-                            await driver.backgroundApp(suspendTime);
-                            await screen.loadedElement(teamTwo.name); // wait for team
+                        if (!shared.preventApplicationCrashCauesByAutomation(driver)) {
+                            if (appSuspendResume) {
+                                await screen.loadedElement(teamTwo.name); // wait for team
+                                await driver.backgroundApp(suspendTime);
+                                await screen.loadedElement(teamTwo.name); // wait for team
+                            }
                         }
-
                         await screen.loadedTeamDetails(teamTwo);
 
                         await shared[`testSomePageNavigated${transition}`](screen);
