+++ e2e/nested-frame-navigation/e2e/layout-root/single-frame.e2e-spec.ts
@@ -1,24 +1,27 @@
 import { AppiumDriver, createDriver, logWarn, nsCapabilities } from "nativescript-dev-appium";
 
-import { Screen, playersData, home, somePage, otherPage, teamsData, driverDefaultWaitTime } from "./screen";
-import * as shared from "./shared.e2e-spec";
-import { suspendTime, appSuspendResume, dontKeepActivities, transitions } from "./config";
-import { TabNavigationScreen } from "./tab-navigation-screen";
+import { Screen, playersData, home, somePage, teamsData, driverDefaultWaitTime } from "../screens/screen";
+import * as shared from "../screens/shared";
+import { suspendTime, appSuspendResume, dontKeepActivities, allTransitions } from "../config";
+import { TabNavigationScreen } from "../screens/tab-navigation-screen";
 
-const rootType = "layout-root";
-describe(rootType, async function () {
+describe("layout-root-with-single-frame", async function () {
     let driver: AppiumDriver;
     let screen: Screen;
+    let transitions = [...allTransitions];
 
     before(async function () {
         nsCapabilities.testReporter.context = this;
         driver = await createDriver();
+        await driver.restartApp();
         screen = new TabNavigationScreen(driver);
-        logWarn("====== layout-root ========");
-        if (dontKeepActivities) {
-            await driver.setDontKeepActivities(true);
+        logWarn("====== layout-root-with-single-frame ========");
+        await driver.setDontKeepActivities(dontKeepActivities);
+        if (shared.isApiLevel19(driver)) {
+            // TODO: known issue https://github.com/NativeScript/NativeScript/issues/6798
+            console.log("Skipping flip transition tests on api level 19");
+            transitions = transitions.filter(tr => !tr.toLowerCase().includes("flip"));
         }
-
         driver.defaultWaitTime = driverDefaultWaitTime;
     });
 
@@ -41,18 +44,11 @@ describe(rootType, async function () {
 
         const playerOne = playersData[`playerOne${transition}`];
         const playerTwo = playersData[`playerTwo${transition}`];
-        const teamOne = teamsData[`teamOne${transition}`];
 
-        describe(`${rootType}-transition-${transition}-scenarios:`, async function () {
+        describe(`layout-root-with-single-frame-transition-${transition}-scenario:`, async function () {
 
             before(async function () {
                 nsCapabilities.testReporter.context = this;
-                if (transition === "Flip" &&
-                    driver.isAndroid && parseInt(driver.platformVersion) === 19) {
-                    // TODO: known issue https://github.com/NativeScript/NativeScript/issues/6798
-                    console.log("skipping flip transition tests on api level 19");
-                    this.skip();
-                }
             });
 
             it("loaded home page", async function () {
@@ -140,176 +136,10 @@ describe(rootType, async function () {
                     await driver.waitForElement(home); // wait for home page
                 }
             });
-
-            it("loaded layout root with multi nested frames", async function () {
-                await screen.navigateToLayoutWithMultiFrame();
-                await screen.loadedLayoutWithMultiFrame();
-            });
-
-            it("loaded players list", async function () {
-                await screen.loadedPlayersList();
-            });
-
-            it("loaded teams list", async function () {
-                await screen.loadedTeamsList();
-            });
-
-            it("loaded player details and go back twice", async function () {
-                await shared.testPlayerNavigated(playerTwo, screen);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(playerTwo.name); // wait for player
-                }
-
-                await shared.testPlayerNavigatedBack(screen, driver);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(playerOne.name); // wait for players list
-                }
-
-                await shared.testPlayerNavigated(playerTwo, screen);
-                await shared.testPlayerNavigatedBack(screen, driver);
-            });
-
-            it("navigate players parent frame and go back", async function () {
-                await shared[`testSomePageNavigated${transition}`](screen);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(somePage); // wait for some page
-                }
-
-                if (driver.isAndroid) {
-                    await driver.navBack(); // some page back navigation
-                } else {
-                    await screen.goBackFromSomePage();
-                }
-
-                await screen.loadedPlayersList();
-            });
-
-            it("loaded players details and navigate parent frame and go back", async function () {
-                await shared.testPlayerNavigated(playerTwo, screen);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(playerTwo.name); // wait for player
-                }
-
-                await shared[`testSomePageNavigated${transition}`](screen);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(somePage); // wait for some page
-                }
-
-                if (driver.isAndroid) {
-                    await driver.navBack(); // some page back navigation
-                } else {
-                    await screen.goBackFromSomePage();
-                }
-
-                await screen.loadedPlayerDetails(playerTwo);
-
-                await screen.goBackToPlayersList();
-                await screen.loadedPlayersList();
-            });
-
-            it("loaded layout root with multi nested frames again", async function () {
-                await screen.loadedLayoutWithMultiFrame();
-            });
-
-            it("loaded players list", async function () {
-                await screen.loadedPlayersList();
-            });
-
-            it("loaded teams list", async function () {
-                await screen.loadedTeamsList();
-            });
-
-            it("mix player and team list actions and go back", async function () {
-                await shared.testPlayerNavigated(playerTwo, screen);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(playerTwo.name); // wait for player
-                }
-
-                await shared[`testOtherPageNavigated${transition}`](screen); // "teams" parent frame navigation
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(otherPage); // wait for other page
-                }
-
-                await screen.loadedPlayerDetails(playerTwo);  // assert no changes in the sibling frame
-
-                if (driver.isAndroid) {
-                    await driver.navBack(); // other page back navigation
-                } else {
-                    await screen.goBackFromOtherPage();
-                }
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(teamOne.name); // wait for teams list
-                }
-
-                await screen.loadedTeamsList();
-                await screen.loadedPlayerDetails(playerTwo);  // assert no changes in the sibling frame
-
-                await shared[`testOtherPageNavigated${transition}`](screen);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(otherPage); // wait for other page
-                }
-
-                await shared[`testSomePageNavigated${transition}`](screen);
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(somePage); // wait for some page
-                }
-
-                await screen.loadedOtherPage(); // assert no changes in the sibling frame
-
-                if (driver.isAndroid) {
-                    await driver.navBack(); // some page back navigation
-                } else {
-                    await screen.goBackFromSomePage();
-                }
-
-                await screen.loadedPlayerDetails(playerTwo);
-
-                await screen.goBackToPlayersList();
-                await screen.loadedPlayersList();
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(playerOne.name); // wait for players list
-                }
-
-                await screen.goBackFromOtherPage();
-
-                if (appSuspendResume) {
-                    await driver.backgroundApp(suspendTime);
-                    await driver.waitForElement(teamOne.name); // wait for team
-                }
-
-                await screen.loadedTeamsList();
-            });
-
-            it("loaded home page again", async function () {
-                await screen.resetToHome();
-                await screen.loadedHome();
-            });
         });
     }
 
-    describe(`${rootType}-players-list-slide-transition with parent frame default transition:`, async function () {
+    describe(`layout-root-with-single-frame-players-list-slide-transition with parent frame default transition:`, async function () {
         const playerOne = playersData["playerOneSlide"];
         const playerTwo = playersData["playerTwoSlide"];
 
@@ -368,7 +198,7 @@ describe(rootType, async function () {
         });
     });
 
-    describe(`${rootType}-players-list-slide-transition with parent frame no transition:`, async function () {
+    describe(`layout-root-with-single-frame-players-list-slide-transition with parent frame no transition:`, async function () {
         const playerOne = playersData["playerOneSlide"];
         const playerTwo = playersData["playerTwoSlide"];
 
@@ -427,7 +257,7 @@ describe(rootType, async function () {
         });
     });
 
-    describe(`${rootType}-players-list-flip-transition with parent frame default transition:`, async function () {
+    describe(`layout-root-with-single-frame-players-list-flip-transition with parent frame default transition:`, async function () {
         const playerOne = playersData["playerOneFlip"];
         const playerTwo = playersData["playerTwoFlip"];
 
@@ -490,14 +320,14 @@ describe(rootType, async function () {
         });
     });
 
-    describe(`${rootType}-players-list-flip-transition with parent frame no transition:`, async function () {
+    describe(`layout-root-with-single-frame-players-list-flip-transition with parent frame no transition:`, async function () {
         const playerOne = playersData["playerOneFlip"];
         const playerTwo = playersData["playerTwoFlip"];
 
         before(async function () {
             nsCapabilities.testReporter.context = this;
         });
-        
+
         it("loaded layout root with nested frames", async function () {
             await screen.navigateToLayoutWithFrame();
             await screen.loadedLayoutWithFrame();
