+++ e2e/nested-frame-navigation/e2e/screens/tabview-navigation-screen.ts
