+++ e2e/nested-frame-navigation/e2e/tab-navigation-root/tabs-root.e2e-spec.ts
@@ -1,26 +1,31 @@
 import { AppiumDriver, createDriver, logWarn, nsCapabilities } from "nativescript-dev-appium";
 
-import { Screen, playersData, teamsData } from "./screen";
-import * as shared from "./shared.e2e-spec";
-import { suspendTime, appSuspendResume, dontKeepActivities, transitions } from "./config";
-import { TabNavigationScreen } from "./tab-navigation-screen";
+import { Screen, playersData, teamsData } from "../screens/screen";
+import * as shared from "../screens/shared";
+import { suspendTime, appSuspendResume, dontKeepActivities, allTransitions } from "../config";
+import { TabNavigationScreen } from "../screens/tab-navigation-screen";
 
 const roots = ["TabsTop", "TabsBottom"];
 
-const rootType = "tabs-root";
-describe(rootType, async function () {
+describe("tab-navigation-tabs-root", async function () {
     let driver: AppiumDriver;
     let screen: Screen;
+    let transitions = [...allTransitions];
 
     before(async function () {
         nsCapabilities.testReporter.context = this;
-        logWarn(`====== ${rootType} ========`);
+        logWarn(`====== "tabs-root" ========`);
         driver = await createDriver();
+        await driver.restartApp();
         screen = new TabNavigationScreen(driver);
-        if (dontKeepActivities) {
-            await driver.setDontKeepActivities(true);
+
+        if (shared.isApiLevel19(driver)) {
+            // TODO: known issue https://github.com/NativeScript/NativeScript/issues/6798
+            console.log("Skipping flip transition tests on api level 19");
+            transitions = transitions.filter(tr => !tr.toLowerCase().includes("flip"));
         }
 
+        await driver.setDontKeepActivities(dontKeepActivities);
         driver.defaultWaitTime = 8000;
     });
 
@@ -40,7 +45,7 @@ describe(rootType, async function () {
 
     for (let index = 0; index < roots.length; index++) {
         const root = roots[index];
-        describe(`${rootType}-${root}-scenarios:`, async function () {
+        describe(`tab-navigation-tabs-root-${root}-scenario:`, async function () {
 
             before(async function () {
                 nsCapabilities.testReporter.context = this;
@@ -54,17 +59,10 @@ describe(rootType, async function () {
                 const teamOne = teamsData[`teamOne${transition}`];
                 const teamTwo = teamsData[`teamTwo${transition}`];
 
-                describe(`${rootType}-${root}-transition-${transition}-scenarios:`, async function () {
+                describe(`tabs-root-${root}-transition-${transition}-scenario:`, async function () {
 
                     before(async function () {
                         nsCapabilities.testReporter.context = this;
-
-                        if (transition === "Flip" &&
-                            driver.isAndroid && parseInt(driver.platformVersion) === 19) {
-                            // TODO: known issue https://github.com/NativeScript/NativeScript/issues/6798
-                            console.log("skipping flip transition tests on api level 19");
-                            this.skip();
-                        }
                     });
 
                     it("loaded home page", async function () {
